refactor(blank): extract a single input component
